add ScrapperQueueProducer, refactor BotClient
